Added a setup_init() that takes care of common initialization functions that all umobj utilities ought to be calling.
This is nice because it lets us centralized these tasks in their own function
without having to update all the scripts in bin/ every time a new
initialization taks comes along.

This commit also adds a custom sys.excepthook to handle uncaught exceptions
in nicer ways without resoting to try/except blocks all over the codebase.
As of right now, this excepthook shows a custom logging message when an
IOError is raised.  This common initialization task was the impetus for
creating umobj_init().