Make short output, listing only key names, the default for lsobj.
Where before the mode and moddate of keys were shown by default, now -l is
needed to expose these attributes.  This is more in line with the behavior of
/bin/ls on UNIX systems.

Also, -l is required for -m and -b to have any visible effect.  This is also
just following the conventions of /bin/ls.